Move scaffold command to cmd package
